requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
jwt==1.3.1
motor==3.3.2
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.24.0.post1
